backend/requirements.txt
torch>=2.2.0
torchvision>=0.17.0
timm>=0.9.0
Pillow>=10.0.0
numpy>=1.26.0

fastapi>=0.115
uvicorn[standard]>=0.32
pillow>=10.4
python-multipart>=0.0.12
pydantic>=2.9